move next types to @firebase/app-types/next
